move flow_* headers into flow subfolder
