+++ benchmark/benchmark_bedrock_channel.cpp
@@ -10,7 +10,7 @@
 #include "xtensor/xrandom.hpp"
 
 #include "fastscapelib/grid/raster_grid.hpp"
-#include "fastscapelib/flow_graph.hpp"
+#include "fastscapelib/flow/flow_graph.hpp"
 #include "fastscapelib/algo/flow_routing.hpp"
 #include "fastscapelib/bedrock_channel.hpp"
