+++ include/fastscapelib/fastscapelib.hpp
@@ -10,9 +10,9 @@
 #include "fastscapelib/version.hpp"
 #include "fastscapelib/bedrock_channel.hpp"
 #include "fastscapelib/algo/flow_routing.hpp"
-#include "fastscapelib/flow_graph.hpp"
-#include "fastscapelib/flow_router.hpp"
-#include "fastscapelib/sink_resolver.hpp"
+#include "fastscapelib/flow/flow_graph.hpp"
+#include "fastscapelib/flow/flow_router.hpp"
+#include "fastscapelib/flow/sink_resolver.hpp"
 #include "fastscapelib/grid/base.hpp"
 #include "fastscapelib/grid/unstructured_mesh.hpp"
 #include "fastscapelib/grid/structured_grid.hpp"
