+++ python/src/flow_graph.hpp
@@ -5,11 +5,11 @@
 
 #include "pybind11/pybind11.h"
 
-#include "fastscapelib/flow_graph.hpp"
-#include "fastscapelib/flow_router.hpp"
-#include "fastscapelib/flow_router_factory.hpp"
-#include "fastscapelib/sink_resolver.hpp"
-#include "fastscapelib/sink_resolver_factory.hpp"
+#include "fastscapelib/flow/flow_graph.hpp"
+#include "fastscapelib/flow/flow_router.hpp"
+#include "fastscapelib/flow/flow_router_factory.hpp"
+#include "fastscapelib/flow/sink_resolver.hpp"
+#include "fastscapelib/flow/sink_resolver_factory.hpp"
 #include "fastscapelib/utils/xtensor_utils.hpp"
 
 #include "flow_router.hpp"
