+++ python/src/flow_router.hpp
@@ -8,9 +8,9 @@
 
 #include "pytensor_utils.hpp"
 
-#include "fastscapelib/flow_router.hpp"
-#include "fastscapelib/flow_router_factory.hpp"
-#include "fastscapelib/flow_graph.hpp"
+#include "fastscapelib/flow/flow_router.hpp"
+#include "fastscapelib/flow/flow_router_factory.hpp"
+#include "fastscapelib/flow/flow_graph.hpp"
 
 
 namespace py = pybind11;
