+++ test/test_flow_graph.cpp
@@ -1,7 +1,7 @@
-#include "fastscapelib/flow_graph.hpp"
-#include "fastscapelib/flow_router.hpp"
-#include "fastscapelib/flow_router_factory.hpp"
-#include "fastscapelib/sink_resolver.hpp"
+#include "fastscapelib/flow/flow_graph.hpp"
+#include "fastscapelib/flow/flow_router.hpp"
+#include "fastscapelib/flow/flow_router_factory.hpp"
+#include "fastscapelib/flow/sink_resolver.hpp"
 #include "fastscapelib/grid/raster_grid.hpp"
 
 #include "xtensor/xtensor.hpp"
